swig: Add support for SWIG 4
* Patch autogenerated esys_bindings.py from setup.py on build to fix
  collision of python builtin `property`
* Use functools.partial to ensure self / cls is not passed to any of the
  TSS functions.

Fixes: #4
